+++ index.js
@@ -1,3 +1,8 @@
+var Promise = require('bluebird');
+Promise.config({
+    cancellation: true, //used for canceling promise chain in case of early request response
+});
+
 exports.CLI              = require('./lib/cli');
 exports.Express          = require('express');
 exports.EXPRESS_VERSION  = parseInt(require('express/package.json').version[0]);
@@ -5,6 +10,7 @@ exports.CouchbaseODM     = require('kouchbase-odm');
 exports.Couchbase        = require('couchbase');
 exports.AppManager       = require('./lib/express/appManager.js');
 exports.App              = require('./lib/express/app.js');
+exports.Response         = require('./lib/express/response.js');
 exports.Router           = require('./lib/express/router.js');
 exports.Route            = require('./lib/express/route.js');
 exports.RequestType      = require('./lib/express/requestType.js');