+++ tests/unit/cli/commands/staticData.js
@@ -67,29 +67,37 @@ describe('`staticData` command', function() {
         });
 
         describe('-r, --refresh option', function() {
-            it('should call provided callback with an Error when staticData has never been loaded (so we cant reload it)', function() {
+            it('should log an Error when staticData has never been loaded (so we cant reload it)', function() {
                 var callback = sinon.spy();
+                var logSpy = sinon.spy();
                 this.staticDataGetLoadOptionsStub.returns();
 
-                return this.action({
+                return this.action.call({
+                    log: logSpy
+                },{
                     options: {refresh: true}
-                }, callback).should.be.rejected.then(function() {
+                }, callback).then(function() {
                     callback.should.be.calledOnce;
-                    callback.should.be.calledWith(sinon.match.string);
+                    logSpy.should.have.been.calledOnce;
+                    logSpy.should.have.been.calledWith(sinon.match.instanceOf(Error));
                 });
             });
 
             it('should call provided callback without any arguments when staticData has been successfully reloaded', function() {
                 var callback = sinon.spy();
+                var logSpy = sinon.spy();
                 this.staticDataGetLoadOptionsStub.returns([
                     {
                         odm: ['/some/path'],
                     }
                 ]);
 
-                return this.action({
+                return this.action.call({
+                    log: logSpy
+                },{
                     options: {refresh: true}
                 }, callback).should.be.fulfilled.then(function() {
+                    logSpy.should.have.been.calledOnce;
                     callback.should.be.calledOnce;
                     callback.should.be.calledWith();
                 });