+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v0.6.0
+
+* [ADDED] common cli initialization procedure via new static `CLI.init` method
+* [ADDED] CLI - ls -r (aka. list routes) add additional column listing applied route's middlewares
+* [ADDED] react on route middleware's fulfillment value. Allows to stop processing a request and respond to the request earlier.
+* [FIXED] CLI - `vantage` remote authentication issues (authentication was made to work but has been disabled for now)
+* [FIXED] min. `bi-config` version dumped to `0.7.1` - the version searches for config file name 'config.json5' by default
+* [FIXED] `staticData.loader` - takes db config from `storage:postgres` config path instead of `storage:sequelize`
+* [FIXED] CLI - support data logging to a remote cli console
+
 ## v0.5.1
 
 * [FIXED] frozen CLI when we execute `integrity` cli command if no app is connected